requirements.txt
mysql-connector-python==8.2.0
protobuf==4.21.12
wxPython==4.2.1